ingest/requirements.txt
pika
mysql-connector